requirements.txt
streamlit
google-generativeai
numpy
pymupdf
pymupdf4llm
arxiv
python-dotenv
